split azblob reader and writer
